Feature: New Layout - Change route
