+++ src/client/assets/styles/styles.css
@@ -125,12 +125,12 @@ input[type=email]:focus, input[type=text]:focus, input[type=password]:focus, but
   position: relative;
   height: 450px;
   width: 200px;
-  top: 200px;
-  left: 400px;
+  top: 100px;
+  left: 100px;
 }
 
 .image-wrapper > img {
-  height: 50%;
+  height: 80%;
   vertical-align: center;
 }
 
@@ -147,11 +147,21 @@ input[type=email]:focus, input[type=text]:focus, input[type=password]:focus, but
   height: 100vh;
   width: 100%;
   background-color: #010101;
-  opacity: 0.5;
+  opacity: 0.7;
+}
+
+.name {
+  font-family: 'Raleway', sans-serif;
+  font-size: 100px;
+}
+
+.name.small-screen {
+  font-family: 'Raleway', sans-serif;
+  font-size: 70px;
 }
 
 .name-slogan > .large-screen {
-  padding: 250px 0 0 0;
+  padding: 150px 0 0 0;
 }
 
 .name-slogan > .small-screen {
@@ -168,18 +178,10 @@ input[type=email]:focus, input[type=text]:focus, input[type=password]:focus, but
 
 .slogan {
   font-family: 'Roboto Condensed', sans-serif;
-  font-size: 24px;
+  font-size: 34px;
 }
 
-.signup-callout {
-  font-family: 'Roboto Condensed', sans-serif;
-  font-size: 50px;
-}
 
-.name {
-  font-family: 'Raleway', sans-serif;
-  font-size: 70px;
-}
 
 .chevron-down {
   position: relative;
@@ -222,6 +224,11 @@ input[type=email]:focus, input[type=text]:focus, input[type=password]:focus, but
 .section-signup {
   background-color: #010101;
   padding: 70px 0;
+}
+
+.signup-callout {
+  font-family: 'Roboto Condensed', sans-serif;
+  font-size: 40px;
 }
 
 .section-footer {
