+++ src/client/components/Homepage/Homepage.jsx
@@ -1,68 +1,77 @@
 import React, { Component } from 'react';
-import { NavigationBar } from '../NavigationBar/NavigationBar.jsx'
-import SignupForm from './SignupForm.jsx';
-import Footer from '../Footer/Footer.jsx';
 
-export default class Homepage extends Component {
+class Layout extends Component {
   render() {
-    return (
-      <div className="row center-block">
-        <div className="space" />
-
-        <div className="col-md-4">
-          <div className="text-center">
-            <img className="hidden-xs hidden-sm" src={require('../../assets/img/africa.png')} alt="africa" height="300" />
-            <div className="pad-10" />
-            <h1 className="shop-name">Afiora</h1>
-            <h3 className="brand-title">Building bridges, communicating values.</h3>
+    return(
+      <div className="container-fluid layout">
+        <div className="row section-banner bg-danger">
+          <div className="banner-overlay"/>
+          <div className="chevron-down text-center">
+            <a href="#mission">
+              <i className="fa fa-chevron-down fa-5" aria-hidden="true"></i>
+            </a>
+          </div>
+          <div className="hidden-sm hidden-xs col-md-6">
+            <div className="image-wrapper">
+              <img src={require('../../assets/img/africa.png')} alt="africa_logo"/>
+            </div>
+          </div>
+          <div className="col-sm-12 col-md-6 name-slogan text-center">
+            <div className="visible-xs visible-sm small-screen"/>
+            <div className="hidden-sm hidden-xs large-screen"/>
+            <h1 className="text-uppercase name small-screen visible-xs visible-sm">Afiora</h1>
+            <h1 className="text-uppercase name hidden-sm hidden-xs">Afiora</h1>
+            <h4 className="slogan">Building bridges and communicating values through e-commerce.</h4>
           </div>
         </div>
-
-        <div className="col-md-4 col-md-offset-4 col-lg-4 col-lg-offset-4">
-          <div className="sign-in-wrapper text-center">
-            <SignupForm />
+        <div className="row section-mission" id="mission">
+          <div className="row text-center"><h2>Afiora Right Now</h2></div>
+          <div className="col-sm-12 col-md-4">
+            <div className="mission-card">
+              <div className="mission-img-holder">
+                <img src={require('../../assets/img/placeholder.jpg')} alt=""/>
+              </div>
+              <div className="mission-desc">
+                <h3>Caption</h3>
+                <p>Description</p>
+              </div>
+            </div>
           </div>
-        </div>
-
-        <div className="container">
-          <div className="row">
-            <div className="space" />
-            <div className="col-md-12 about">
-              <div className="col-sm-12">boo</div>
-              <h3 className="text-center">About Us</h3>
-              <div className="row">
-                <div className="col-md-6">
-                  <h4>Vision</h4>
-                  <p>To be the one stop shop for everything made in Africa.</p>
-                </div>
-                <div className="col-md-6">
-                  <h4>Mission</h4>
-                  <p>Building bridges and communicating African values in form
-                    of product and services to the world through E-commerce.
-                  </p>
-                </div>
-                <div className="col-md-6">
-                  <h4>Belief</h4>
-                  <p>We believe that Africa the birthplace of humanity will one day
-                    be the world’s largest economy and we have a major role to play in that.
-                  </p>
-                </div>
-                <div className="col-md-6">
-                  <h4>Afiora Today</h4>
-                  <ul>
-                    <li>The first African indigenous mall</li>
-                    <li>Products available from African countries</li>
-                    <li>Shoppers favourite for everything made in Africa</li>
-                    <li>We sell only Authentic products</li>
-                    <li>Unbeatable price</li>
-                  </ul>
-                </div>
+          <div className="col-sm-12 col-md-4">
+            <div className="mission-card">
+              <div className="mission-img-holder">
+                <img src={require('../../assets/img/placeholder.jpg')} alt="mission_img"/>
+              </div>
+              <div className="mission-desc">
+                <h3>Caption</h3>
+                <p>Description</p>
+              </div>
+            </div>
+          </div>
+          <div className="col-sm-12 col-md-4">
+            <div className="mission-card">
+              <div className="mission-img-holder">
+                <img src={require('../../assets/img/placeholder.jpg')} alt=""/>
+              </div>
+              <div className="mission-desc">
+                <h3>Caption</h3>
+                <p>Description</p>
               </div>
             </div>
           </div>
         </div>
-        <Footer />
+        <div className="row section-signup text-center">
+          <div className="col-sm-4 col-sm-offset-4 col-md-4 col-md-offset-4">
+            <h2 className="text-uppercase signup-callout">Become a merchant</h2>
+            <button className="btn signup">Signup</button>
+          </div>
+        </div>
+        <div className="row section-footer text-center">
+          <p>&copy; 2017 Afiora Ltd.</p>
+        </div>
       </div>
     );
   }
 }
+
+export default Layout;
